Changed to http server and made output html with color formatting #39

+++ PythonTester/Main.py
@@ -43,7 +43,7 @@ parser.add_argument('-p', '--print', dest='print', action='store_true', help='Pr
 parser.add_argument('-e', '--error', dest='error', action='store_true', help='Print the mutated code\'s errors to the console')
 parser.add_argument('-r', '--report', dest='report', action='store_true', help='Generate mutation report file')
 parser.add_argument('-m', '--modify', dest='modify', action='store_true', help='Attempt to login to change list of mutations')
-parser.add_argument('--timeout', dest='timeout', type=int, default=None, help='Optional global timeout (in seconds) for each file\'s mutation loop. Files can override via config.')
+parser.add_argument('--timeout', dest='timeout', type=float, default=None, help='Optional global timeout (in seconds) for each file\'s mutation loop. Files can override via config.')
 
 def generate_default_config():
     DEFAULT_CONFIG_TEXT = """\
